Update README.md
These references are being removed related to the recent discovery of plagiarism on Siraj's channel. This follows a pattern of little to no crediting of sources in code examples and videos.

For more background:
https://twitter.com/AndrewM_Webb/status/1183150368945049605

I don't feel I can confidently sent viewers in the direction of his content at this time. I would be happy to reconsider if and when it becomes appropriate to do so in the future. 

😞